Lots of stuff changed
Created a growable nn.Module that allows for growing a matrix and giving
each section a different learning rate.  This is done by breaking the
matrix up, which I'll have to screw with later.  This is tested, and
allows for the expected behavior.
Created the basic structure of the transformer model we will be using.
Still need RoPE, and some initialization stuff for the growable module.
split out some stuff into types.py for cleanliness.